requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
itsdangerous==2.2.0
Werkzeug==3.0.3
requests==2.32.3
